Convert settings.js to TypeScript
